remove usage of formfieldgroupsloop because of issues with css grid and conditionals
